+++ CHANGELOG.md
@@ -11,8 +11,7 @@
 - 废除 `f-alert` 组件 `icon` 属性
 - 新增 `f-alert` 组件 `prefixIcon` 和 `closeIcon` 插槽
 - 新增 `f-close-btn` 组件 `color` 属性
-- 废除 `f-button` 组件 `before-icon` `after-icon` 配置项
-- 新增 `f-button` 组件 `beforeIcon` `afterIcon` `loadingIcon` 插槽
+- `f-button` 组件 `before-icon` `after-icon` `loading-icon` 配置项改为 `VNode` 类型，默认值修改为 `null`
 - 废除 `f-ink` 组件 `before-icon` `after-icon` 配置项
 - 新增 `f-ink` 组件 `beforeIcon` `afterIcon` 插槽
 - 废除 `f-avatar` 组件 `icon` 配置项